requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
openai>=1.58.1,<2.0.0
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
email-validator
pytest
python-multipart